Red Neuronal/requirements.txt
tensorflow-gpu==2.10.0
opencv-python==4.10.0
numpy==1.26.4
cudatoolkit==11.2
cudnn==8.1.0
